+++ dhis-2/dhis-web/dhis-web-caseentry/src/main/java/org/hisp/dhis/caseentry/screen/DefaultDataEntryScreenManager.java
@@ -179,10 +179,10 @@ public class DefaultDataEntryScreenManager
         return calculatedDataElementIds;
     }
 
-    public Map<CalculatedDataElement, Map<DataElement, Integer>> getNonSavedCalculatedDataElements(
+    public Map<CalculatedDataElement, Map<DataElement, Double>> getNonSavedCalculatedDataElements(
         ProgramStage programStage )
     {
-        Map<CalculatedDataElement, Map<DataElement, Integer>> calculatedDataElementMap = new HashMap<CalculatedDataElement, Map<DataElement, Integer>>();
+        Map<CalculatedDataElement, Map<DataElement, Double>> calculatedDataElementMap = new HashMap<CalculatedDataElement, Map<DataElement, Double>>();
 
         Collection<DataElement> dataElements = programStageDataElementService.getListDataElement( programStage );
