+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-datadictionary/src/main/java/org/hisp/dhis/dd/action/dataelement/ShowUpdateDataElementFormAction.java
@@ -111,9 +111,9 @@ public class ShowUpdateDataElementFormAction
         return dataElementGroups;
     }
 
-    private Map<String, Integer> factorMap;
+    private Map<String, Double> factorMap;
 
-    public Map<String, Integer> getFactorMap()
+    public Map<String, Double> getFactorMap()
     {
         return factorMap;
     }
